Try to handle duplicate tags better
If, for example, a single source file contains multiple private blocks
which each have otwherwise identical labels in, we would previously
generate multiple identical tag entries for them, using a search to
locate the label. We now attempt to detect this and generate line
number-based tags for the second and subsequent definitions of the
label, which means things like tselect can show them all.